extracting `storage` parsing logic to `storage.rs`
